Accidentally included removed test code
This was kind of going nowhere. There’s not really a good way to do
this… in any case, you end up just going through the valid types’ `+`
functions. There’s no “plussable” protocol, and even if there were, it
would have a Self reference so you can’t cast to it.